requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
qdrant-client
langchain
openai
